Change sample app to support ios 11
