+++ SampleApp/Sources/AppDelegate.swift
@@ -2,15 +2,20 @@ import UIKit
 
 @UIApplicationMain
 class AppDelegate: UIResponder, UIApplicationDelegate {
+
+    var window: UIWindow?
+
     func application(_ application: UIApplication,
                      didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
-        return true
-    }
 
-    // MARK: UISceneSession Lifecycle
-    func application(_ application: UIApplication,
-                     configurationForConnecting connectingSceneSession: UISceneSession,
-                     options: UIScene.ConnectionOptions) -> UISceneConfiguration {
-        return UISceneConfiguration(name: "Default Configuration", sessionRole: connectingSceneSession.role)
+        let navigationController = UINavigationController(rootViewController: MainViewController())
+        let window = UIWindow()
+        window.backgroundColor = .white
+        window.rootViewController = navigationController
+        window.makeKeyAndVisible()
+
+        self.window = window
+
+        return true
     }
 }
