--- SampleApp/Sources/SceneDelegate.swift
@@ -1,19 +1,0 @@
-import UIKit
-
-class SceneDelegate: UIResponder, UIWindowSceneDelegate {
-    var window: UIWindow?
-
-    func scene(_ scene: UIScene,
-               willConnectTo session: UISceneSession,
-               options connectionOptions: UIScene.ConnectionOptions) {
-        if let windowScene = scene as? UIWindowScene {
-            let navigationController = UINavigationController(rootViewController: MainViewController())
-
-            let window = UIWindow(windowScene: windowScene)
-            window.backgroundColor = .white
-            window.rootViewController = navigationController
-            self.window = window
-            window.makeKeyAndVisible()
-        }
-    }
-}
